refactor: move o middleware de enriquecimento do correlationId para ficar junto do método de config dos logs
